samples: bluetooth: add BLE throughput sample
This commit adds a sample to measure BLE throughput.
The sample has to be flashed on two boards that connect
to each other automatically. The test is started via a
key press on the console of one of the two boards.
At the end of the test, the sample will print the
metrics as measured locally and by the peer.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>