samples: nrf_desktop: Load settings after GATT services setup
Prior to this change, CCC value for HIDS service was loaded from flash
before initialization of the service and was not written to GATT
database.

This fixes DESK-288 in case if there was BLE disconnection before
peripheral reset. For other cases, a subsequent PR to Zephyr is going to
be prepared.

Jira: DESK-288

Signed-off-by: Filip Kubicz <filip.kubicz@nordicsemi.no>