samples: Add a LTE-BLE gateway sample
Add a sample to showcase a LTE-BLE gateway.
The application connects to the nRF52840 on the pca10090 board,
which is running the BLE HCI controller over UART, and scans for
a Thingy52. Once it established a connection with the Thingy52,
it subscribes to its orientation characteristic and starts to
collect orientation data from the Thingy52 and simulated GPS data
generated locally. Whenever the Thingy52 is flipped upside down the
gateway sends sends all collected data to nRF Cloud
via LTE network connection.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>
Signed-off-by: Rune Holmgren <rune.holmgren@nordicsemi.no>
Signed-off-by: Even Falch-Larsen <even.falch-larsen@nordicsemi.no>