samples: nrf9160: asset_tracker: Make FLIP trigger configurable
This patch adds an option to Kconfig to set up a button or switch
to trigger polling of simulated accelerometer data and send
the resulting FLIP data to nRF Cloud.

Also introduces check that sending of sensor data is enabled
before trying to poll accelerometer and calculate orientation.

Function flip_poll is renamed to flip_send.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>