subsys: esb: Fix stopping Enhanced ShockBurst rx during acking
In the current implementation it's possible to call stop rx in the small
time window where the device is sending an ack package. This will cause
the stop rx to fail, and if the application isn't set up to handle this
error with a retry it's going to be a problem. This commit adds support
for disabling rx in this time window.

Signed-off-by: Rune Holmgren <Rune.Holmgren@nordicsemi.no>