lib: gps_sim: Make position update time configurable
The GPS simulator generates positions based on the system uptime,
resulting in possibly new position updates every millisecond.
This patch makes it possible to configure how often the position
will be updated to be a but more like a real GPS that gets a new
position fix at certain intervals.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>