samples: nrf9160: asset_tracker: Add common handling of sensors
This patch does some groundwork for more easily adding new
sensor types to the sample. Now available sensors are placed
in an array which is looped over and attached to nRF Cloud.
All sensors have to be confirmed attached to nRF Cloud before
they are initialized and sensor data can be sent.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>