ci: Fix CI builds on branches
It does not make sense to tell the compliance script to report issues
as a comment, when the checked commit is inside a branch or on master.
This commit only sets these extra arguments if it's a pull request.
Fixes build status on 'master' and similar branches that is triggered
by merging.

Signed-off-by: Ulrich Myhre <ulrich.solli.myhre@nordicsemi.no>