event_manager: Code cleanup
Commit cleans up the code. There are also to issues fixed.
First is the possibility to compute invalid pointer to the container.
This issue should not be seen as node is the first element of the
event header.
Second is the possibility to access event after being freed. This could
happen if event processor picked up event right after adding it to
the list but before doing submission profiling. Should also not
happen as it would require specific scheduling conditions.

Jira:DESK-417

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>