subsys: bootloader: c_runtime_setup: Add missing function for ISR
In zephyr they have added a new ISR function to the interrupt list
this function need to be defined for the code to be compiled in but we
are not going to use it. So we just alias it to the same
`dummy_handler` as the rest of the ISRs.

Also since we are currently shareing Kconfig with the application we
need to alias the assert post action function for running tests.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>