subsys: bootloader: Fix secure boot for Kconfig being moved after DTC
Add symbols which previously were defined by using device tree and
Kconfig into a seperate headerfile called `bootloader.h`. This file
contains `DT_FLASH_AREA_*_OFFSET/SIZE` defines used by the bootloader.
This is a fix to get secure boot to build with master and is a stepping
stone until multi-image builds are supported.

Add flash offset to the kconfig for the application so that the
application is placed in the correct flash address based on what is
described in `bootloader.h`.

By having all the defines being prefixed with `DT_` we have
compatibility with the coming multi-image build support. Also by having
all the defines in a `bootloader.h` file we can easily remove the file
and remove the header file include where the compiler complains/breaks
when support is added.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>