app: asset_tracker: Add option to not include modem info.
As current application mandates CONFIG_MODEM_INFO to be
enabled this makes the use of MODEM_INFO optional.

main.c has be to aware of the underlying modules being
included or not and have to be testing with ifdef's if
relevant code should be included based on the CONFIG_MODEM_INFO
define.

Signed-off-by: Glenn Ruben Bakke <glenn.ruben.bakke@nordicsemi.no>