samples: cmake: Fix issue where multi-image didn't compile on windows
Since paths created by the build system are longer than 259 characters
on windows. Building any thing with multi-image is problematic this is 1
of 3 optimalizations which can be done to reduce the path lengths.

The two others are 1. Use relative paths to where gcc is invoked instead
of absolute paths and 2. is not including file path in build directories
(location of file) this seems to be a part of CMake and may be harder to
change.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>