subsys: partition_manager: Fix delimiter for split on `:` to `|`
The problem is that a split on `:` in windows would result in splitting
`c:\` on windows which would create invalid paths. Windows does not
allow naming a folder `|` linux does. We should create a limitation in
the build system to check for this as any folder with characters which
are forbidden in Windows would break building on windows. Hence this
would be safe to assume is a good enough character for splitting.

This also adds a check for if a path contains `|` to issue an error
message to the user that you are not allowed to have paths with `|`.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>