scripts: partition_manager: add span type 'one_of'
Support new type of 'span' which only spans one partition.
This is useful when defining spans which will span over different
partitions depending on what images are configured in.

One example of this is the S0 span of B0 where either MCUboot or
the root image app can be placed depending on whether or not MCUboot
is included in the build.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>