lib: bsdlib: assert on untranslated errno
When an unknown errno is received from bsdlib, bsd_os.c translates
the errno to EINVAL, effectively hiding it. This is confusing,
since EINVAL is a very generic errno that can be returned from
a myriad of different places in the code.

With this patch, any untranslated errno set by bsdlib will trigger
an assert to alert the developer of the missing translation code
in bsd_os.c. In case asserts are disabled, the untranslated errno
will be logged on the highest possible level and replaced with
a magic value to make this situation visible to the developer.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>