applications: nrf_desktop: Avoid spinlock in button hw interface
GPIO callback control is rather expensive operation, especially if
it has to be done for all rows. We need to avoid locking interrupts
while doing it as it can lead to problems with BLE.
Code defers GPIO IRQ to work and makes sure handler is canceled when
needed. State is safe as it is touched only from system workqueue
context.

Jira:DESK-602

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>