lib: modem_info: Fix RSRP notification handler
RSRP notification handler parsed the CESQ notification incorrectly.

Instead of parsing the CESQ notification in correct format and returning
the value from the parsed values, it sent the AT CESQ command and tried
to parse the response, which has a different format than the
notificaiton. In result, incorrect field was used to fill the
`param_value` variable.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>