drivers: bh1749: Fix DT label once again
This is a follow-up to commit 97a4ca0b872d4b6ab6f5fb4ee2849df792471d51.

The currently used label contains the slave address of the sensor and
the base address of the TWIM peripheral, so it's not a proper one to be
used in the code that is supposed to be generic. This commit changes it
to DT_INST_0_ROHM_BH1749_LABEL.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>