samples: bluetooth: central: enabled bond storage
Enabled bond storage in the Bluetooth Central samples.

Changed the way how the discovery is started for central devices. Now,
the discovery is triggered at the end of the following sequence:

connected -> bt_conn_set_security -> security_changed -> gatt_discover

This assumes that security_changed callback is called regardless of
whether the encryption/pairing has succeeded or not.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>