Bluetooth: Services: NUS: Fix bt_gatt_nus_send with given conn
Fix is_notification_enabled when checking if notifications is enabled
for a given connection. This function used the wrong user_data
definition and compared addresses using the wrong pointer.
The function was also given the wrong attributed.
Also checking if notifications is enabled would fail if indications was
also enabled.
Also deleted the cfg_changed as it only contains information about the
the attribute and the value, not which peer subscribed.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>