samples: lwm2m_client: Update DNS in project configuration
Following the upstream changes, it is no longer needed to enable
DNS_RESOLVER in order to run DNS in the lwm2m library. As it is no
longer used, native IP stack can now be disabled as well.

As we support IPv4 in this sample, we need to specify it by a proper
NET_IPV4/6 configuration in the project.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>