applications: nrf_desktop: Workaround for USB callback race
Due to how USB stack is constructed in Zephyr, driver may issue
callbacks indicating the USB state before application had a chance
to register its callback function.

So far the problem is visible only with Nordic LL, only when settings
are erased and only if device boots with USB connected. In such case
application misses callback about the device being powered.

To fix the problem take the assumption that the device was powered
if we receive suspend event from the driver.

Jira:DESK-633

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>