drivers: hw_cc310: Driver will default be deselected when using mcuboot
When building multi image including mcuboot then mcuboot is compiled
with hard float, however the flags CONFIG_FLOAT and CONFIG_FP_HARDABI
are not set which causes wrong linking.
Therefore hw cc310 will default to yes in mcuboot until this has been
corrected.
This is possible, as cc310 is currently not required in mcuboot.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>