net: lib: Erase pending may cause the modem DFU operation to stall
Doing a modem DFU may cause the modem's erase operation to hang on
`DFU_ERASE_PENDING` this is due to the modem being too busy to do the
actual erase. Either reboot the device or turn of the LTE link until
the erase operation is complete to mitigate this issue.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>